Remove globals.c
