+++ src/av/CMakeLists.txt
@@ -3,7 +3,6 @@ project(utoxAV LANGUAGES C)
 add_library(utoxAV STATIC
     utox_av.c
     audio.c
-    globals.c
     video.c
     filter_audio.c
     )
