+++ src/theme_tables.c
@@ -1,107 +1,7 @@
-#include "friend.h"
-#include "messages.h"
-#include "self.h"
-#include "theme.h"
 #include "theme_tables.h"
-#include "tox.h"
-#include "ui.h"
 
-/* Globals */
+#include <stddef.h> /* NULL */
 
-/* friend.h */
-uint8_t addfriend_status;
-
-/* messages.h */
-pthread_mutex_t messages_lock;
-
-/* self.h */
-struct utox_self self;
-
-/* theme.h */
-uint32_t COLOR_BKGRND_MAIN;
-uint32_t COLOR_BKGRND_ALT;
-uint32_t COLOR_BKGRND_AUX;
-uint32_t COLOR_BKGRND_MENU;
-uint32_t COLOR_BKGRND_MENU_HOVER;
-uint32_t COLOR_BKGRND_MENU_ACTIVE;
-uint32_t COLOR_BKGRND_LIST;
-uint32_t COLOR_BKGRND_LIST_HOVER;
-
-uint32_t COLOR_MAIN_TEXT;
-uint32_t COLOR_MAIN_TEXT_CHAT;
-uint32_t COLOR_MAIN_TEXT_SUBTEXT;
-uint32_t COLOR_MAIN_TEXT_ACTION;
-uint32_t COLOR_MAIN_TEXT_QUOTE;
-uint32_t COLOR_MAIN_TEXT_RED;
-uint32_t COLOR_MAIN_TEXT_URL;
-uint32_t COLOR_MAIN_TEXT_HINT;
-
-uint32_t COLOR_MSG_USER;
-uint32_t COLOR_MSG_USER_PEND;
-uint32_t COLOR_MSG_USER_ERROR;
-uint32_t COLOR_MSG_CONTACT;
-
-uint32_t COLOR_MENU_TEXT;
-uint32_t COLOR_MENU_TEXT_SUBTEXT;
-uint32_t COLOR_MENU_TEXT_ACTIVE;
-
-uint32_t COLOR_LIST_TEXT;
-uint32_t COLOR_LIST_TEXT_SUBTEXT;
-
-uint32_t COLOR_AUX_EDGE_NORMAL;
-uint32_t COLOR_AUX_EDGE_HOVER;
-uint32_t COLOR_AUX_EDGE_ACTIVE;
-uint32_t COLOR_AUX_TEXT;
-uint32_t COLOR_AUX_ACTIVEOPTION_BKGRND;
-uint32_t COLOR_AUX_ACTIVEOPTION_TEXT;
-
-uint32_t COLOR_GROUP_SELF;
-uint32_t COLOR_GROUP_PEER;
-uint32_t COLOR_GROUP_AUDIO;
-uint32_t COLOR_GROUP_MUTED;
-
-uint32_t COLOR_SELECTION_BACKGROUND;
-uint32_t COLOR_SELECTION_TEXT;
-
-uint32_t COLOR_EDGE_NORMAL;
-uint32_t COLOR_EDGE_ACTIVE;
-uint32_t COLOR_EDGE_HOVER;
-
-uint32_t COLOR_ACTIVEOPTION_BKGRND;
-uint32_t COLOR_ACTIVEOPTION_TEXT;
-
-uint32_t COLOR_STATUS_ONLINE;
-uint32_t COLOR_STATUS_AWAY;
-
-uint32_t COLOR_STATUS_BUSY;
-uint32_t COLOR_BTN_SUCCESS_BKGRND;
-uint32_t COLOR_BTN_SUCCESS_TEXT;
-uint32_t COLOR_BTN_SUCCESS_BKGRND_HOVER;
-uint32_t COLOR_BTN_SUCCESS_TEXT_HOVER;
-
-uint32_t COLOR_BTN_WARNING_BKGRND;
-uint32_t COLOR_BTN_WARNING_TEXT;
-uint32_t COLOR_BTN_WARNING_BKGRND_HOVER;
-uint32_t COLOR_BTN_WARNING_TEXT_HOVER;
-
-uint32_t COLOR_BTN_DANGER_BACKGROUND;
-uint32_t COLOR_BTN_DANGER_TEXT;
-uint32_t COLOR_BTN_DANGER_BKGRND_HOVER;
-uint32_t COLOR_BTN_DANGER_TEXT_HOVER;
-
-uint32_t COLOR_BTN_DISABLED_BKGRND;
-uint32_t COLOR_BTN_DISABLED_TEXT;
-uint32_t COLOR_BTN_DISABLED_BKGRND_HOVER;
-uint32_t COLOR_BTN_DISABLED_TRANSFER;
-
-uint32_t COLOR_BTN_INPROGRESS_BKGRND;
-uint32_t COLOR_BTN_INPROGRESS_TEXT;
-uint32_t COLOR_BTN_DISABLED_FORGRND;
-uint32_t COLOR_BTN_INPROGRESS_FORGRND;
-
-uint32_t status_color[4];
-
-/* theme_tables.h */
 const char *COLOUR_NAME_TABLE[] = { "MAIN_BACKGROUND",
                                     "ALT_BACKGROUND",
                                     "MAIN_TEXT",
@@ -176,6 +76,7 @@ const char *COLOUR_NAME_TABLE[] = { "MAIN_BACKGROUND",
                                     "TRANSFER_PROGRESS_OVERLAY_ACTIVE",
                                     "BUTTON_INPROGRESS_TEXT",
                                     NULL };
+
 uint32_t *COLOUR_POINTER_TABLE[] = { &COLOR_BKGRND_MAIN,
                                      &COLOR_BKGRND_ALT,
                                      &COLOR_MAIN_TEXT,
@@ -250,22 +151,3 @@ uint32_t *COLOUR_POINTER_TABLE[] = { &COLOR_BKGRND_MAIN,
                                      &COLOR_BTN_INPROGRESS_FORGRND,
                                      &COLOR_BTN_INPROGRESS_TEXT,
                                      NULL };
-
-/* tox.h */
-UTOX_TOX_THREAD_INIT tox_thread_init;
-
-TOX_MSG       tox_msg, audio_msg, toxav_msg;
-volatile bool tox_thread_msg, audio_thread_msg, video_thread_msg;
-
-bool tox_connected;
-char proxy_address[256]; /* Magic Number inside toxcore */
-
-/* ui.h */
-struct utox_mouse mouse;
-
-uint8_t cursor;
-bool mdown;
-
-char search_data[1024]; // TODO this is NOT where this belongs
-
-double ui_scale;
