+++ src/ui/draw.c
@@ -1,6 +1,3 @@
 #include "draw.h"
 
-/* Globals */
-
-/* draw.h */
 int font_small_lineheight, font_msg_lineheight;
